+++ src/pydiverse/pipedag/materialize/cache.py
@@ -57,6 +57,7 @@ class TaskCacheInfo:
             version=self._task.version,
             timestamp=datetime.now(),
             run_id=ctx.run_id,
+            position_hash=self._task.position_hash,
             input_hash=self._input_hash,
             cache_fn_hash=self._cache_fn_hash,
             output_json=output_json,